refactor: move document and policy compilers to authly-domain
