+++ Cargo.toml
@@ -55,20 +55,16 @@ indoc = "2"
 int-enum = "1"
 itertools = "0.14"
 hostname = "0.4"
-humantime = "2"
 hyper = { version = "1", default-features = false }
 hyper-util = { version = "0.1", features = ["tokio", "server", "http2"] }
 jsonwebtoken = "9"
 k8s-openapi = { version = "0.24.0", features = ["latest"] }
 kube = "0.98"
-maud = { version = "0.26", features = ["axum"] }
 mimalloc = "0.1.43"
 num-derive = "0.4"
 num-traits = "0.2"
 openraft = { version = "0.9", default-features = false }
 pem = "3"
-pest = "2"
-pest_derive = "2"
 postcard = { version = "1", default-features = false, features = ["alloc"] }
 prost-types = "0.13"
 rand = "0.8"
