+++ lib/authly-domain/src/document/error.rs
