+++ lib/authly-domain/src/document/mod.rs
@@ -1,5 +1,3 @@
 pub mod compiled_document;
 pub mod doc_compiler;
 pub mod error;
-
-pub(crate) mod load;
