+++ lib/authly-domain/src/policy/compiler/codegen.rs
@@ -2,7 +2,8 @@ use authly_common::{
     id::{AttrId, EntityId, PropId},
     policy::code::OpCode,
 };
-use authly_domain::id::BuiltinProp;
+
+use crate::id::BuiltinProp;
 
 use super::expr::{Expr, Global, Term};
