+++ lib/authly-domain/src/policy/test_compile.rs
@@ -6,14 +6,14 @@ use authly_common::{
     id::{kind::Kind, AttrId, PropId, ServiceId},
     policy::code::OpCode,
 };
-use authly_domain::id::BuiltinProp;
 
 use crate::{
-    db::service_db::PropertyKind,
     document::{
         compiled_document::{CompiledAttribute, CompiledDocumentData, CompiledProperty},
         doc_compiler::{NamespaceEntry, NamespaceKind, Namespaces},
     },
+    id::BuiltinProp,
+    repo::service_repo::PropertyKind,
 };
 
 const SVC: ServiceId = ServiceId::from_uint(42);
