+++ lib/authly-domain/src/repo/document_repo.rs
@@ -2,16 +2,13 @@ use std::{borrow::Cow, ops::Range};
 
 use authly_common::id::{AnyId, AttrId, DirectoryId, PolicyId, PropId, ServiceId};
 use authly_db::{literal::Literal, param::ToBlob, params, Db, DbError};
-use authly_domain::{
-    audit::Actor,
-    encryption::{DecryptedDeks, EncryptedObjIdent},
-};
 use indoc::indoc;
 use itertools::Itertools;
 use serde_spanned::Spanned;
 use tracing::info;
 
 use crate::{
+    audit::Actor,
     document::{
         compiled_document::{
             CompiledDocument, CompiledEntityAttributeAssignment, CompiledEntityRelation,
@@ -19,10 +16,10 @@ use crate::{
         },
         error::DocError,
     },
+    encryption::{DecryptedDeks, EncryptedObjIdent},
+    repo::{service_repo::PropertyKind, Identified},
     settings::Setting,
 };
-
-use super::{service_db::PropertyKind, Identified};
 
 #[derive(thiserror::Error, Debug)]
 pub enum DocumentDbTxnError {
