+++ lib/authly-domain/src/repo/service_repo.rs
@@ -5,10 +5,11 @@ use authly_common::{
     service::NamespacePropertyMapping,
 };
 use authly_db::{param::ToBlob, params, Db, DbResult, FromRow, Row, TryFromRow};
-use authly_domain::{builtins::Builtins, id::BuiltinProp};
 use indoc::{formatdoc, indoc};
 use serde::{Deserialize, Serialize};
 use tracing::warn;
+
+use crate::{builtins::Builtins, id::BuiltinProp};
 
 #[derive(Debug)]
 pub struct NamespaceProperty {
