+++ lib/authly-domain/src/repo/settings_repo.rs
@@ -1,9 +1,11 @@
 use std::borrow::Cow;
 
 use authly_db::{params, Db, DbResult, Row, TryFromRow};
-use authly_domain::directory::DirKey;
 
-use crate::settings::{Setting, Settings};
+use crate::{
+    directory::DirKey,
+    settings::{Setting, Settings},
+};
 
 struct LocalSetting {
     #[expect(unused)]
