+++ src/util/mod.rs
@@ -1,4 +1,3 @@
-pub mod error;
 pub mod protocol_router;
 pub mod remote_addr;
 pub mod serde;
